Final Modification in CartShopping and modifing Product adding GetHash Equals Override

+++ Models/Cart.cs
@@ -12,7 +12,7 @@ namespace Cowrk_Space_Mangment_System.Models
         [Display(Name ="Client")]
         public bool IsClient { get; set; }
         [Display(Name ="Price For Your Catering")]
-        public int TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
         [Display(Name ="Invoice Date")]
         public DateTime Date { get; set; }
         // Paid or not
